requirements.txt
langchain
langchain-community
langchain-google-genai
google-generativeai
python-dotenv
SpeechRecognition
gTTS
pyaudio
